refactor(domain): extract collectBoardNodeSummaries into board subfolder
Move collectBoardNodeSummaries from useBoardStore.ts to
src/core/domain/board/collectBoardNodeSummaries.ts where it belongs
as a board-level projection concern.
